booking-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
grpcio==1.59.3
grpcio-tools==1.59.3
pydantic==2.5.0
pydantic-settings==2.1.0
